[ADEV-55] Ajouter un bandeau "Proposé par Osmose"
