+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
-import { FC, ReactNode } from 'react';
-import { BrowserRouter as Router, Link, Redirect, Route, Switch } from 'react-router-dom';
+import { FC } from 'react';
+import { BrowserRouter as Router, Redirect, Route, Switch } from 'react-router-dom';
 
 
 import { Login } from "@/view/login.page.tsx";
@@ -13,8 +13,7 @@ import { AudioAnnotator } from "@/view/audio-annotator/audio-annotator.page.tsx"
 import { CreateCampaign } from "@/view/create-campaign/create-campaign.page";
 
 import { StaffOnlyRoute } from "@/routes/staff-only";
-
-import { useAuthService } from "@/services/auth";
+import { AploseSkeleton } from "@/view/global-components/skeleton/skeleton.component.tsx";
 
 import './css/fontawesome/css/fontawesome-5.15.4.min.css';
 import './css/fontawesome/css/solid.min.css'
@@ -24,7 +23,7 @@ import '@ionic/react/css/core.css';
 import './css/ionic-override.css';
 import './css/app.css';
 
-import { IonApp, IonButton, setupIonicReact } from '@ionic/react';
+import { IonApp, setupIonicReact } from '@ionic/react';
 
 import { Provider } from "react-redux";
 import { AppStore } from "@/slices/app.ts";
@@ -34,35 +33,6 @@ setupIonicReact({
   spinner: 'crescent',
 });
 
-
-const AploseSkeleton: FC<{ children?: ReactNode }> = ({ children }) => {
-  const auth = useAuthService();
-  return (
-    <div className="px-5 mx-5">
-      <div className="row text-center">
-        <div className="col-sm-12"><h1>APLOSE</h1></div>
-      </div>
-      <div className="row text-left h-100 main">
-        <div className="col-sm-2 border rounded">
-          <ul>
-            <li><a href="/..">Back to main site</a></li>
-            <li><Link to="/datasets">Datasets</Link></li>
-            <li><Link to="/annotation-campaigns">Annotation campaigns</Link></li>
-            <br/>
-            <li>
-              <IonButton color={ "medium" }
-                         onClick={ auth.logout.bind(auth) }>
-                Logout
-              </IonButton>
-            </li>
-          </ul>
-        </div>
-
-        { children }
-      </div>
-    </div>
-  )
-}
 
 export const App: FC = () => {
   return (
