+++ frontend/src/view/create-campaign/create-campaign.page.tsx
@@ -51,8 +51,7 @@ export const CreateCampaign: React.FC = () => {
 
   return (
     <form id="create-campaign-form"
-          onSubmit={ handleSubmit }
-          className="col-sm-9 border rounded">
+          onSubmit={ handleSubmit }>
       <h1>Create Annotation Campaign</h1>
 
       <GlobalInfoBloc/>
