+++ frontend/src/view/create-campaign/create-edit-campaign.css
@@ -1,11 +1,11 @@
 #create-campaign-form {
-  display: grid;
-  justify-items: center;
-  align-content: baseline;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   gap: 4rem;
   color: var(--ion-color-secondary);
   font-family: var(--ion-font-family);
-  padding: 4rem 2rem;
+  padding: 0 2rem;
 }
 #create-campaign-form > * {
   width: min(100%, 40rem);
