+++ frontend/src/view/global-components/osmose-bar/osmose-bar.component.css
@@ -1,0 +1,25 @@
+#osmose-bar-component,
+#osmose-bar-component #osmose-link {
+  display: grid;
+  grid-auto-flow: column;
+  justify-content: center;
+  align-items: center;
+  font-family: var(--ion-font-family);
+  color: var(--ion-text-color);
+}
+#osmose-bar-component {
+  gap: 0.5rem;
+}
+#osmose-bar-component p {
+  opacity: 0.8;
+}
+#osmose-bar-component #osmose-link {
+  font-weight: 600;
+  letter-spacing: 2px;
+  text-decoration: none;
+}
+#osmose-bar-component img {
+  width: 2rem;
+  height: 2rem;
+  object-fit: contain;
+}
