+++ frontend/src/view/global-components/skeleton/skeleton.component.css
@@ -1,0 +1,26 @@
+#aplose-skeleton-component {
+  display: grid;
+  grid-template-rows: auto 1fr auto;
+  grid-template-columns: 200px 1fr;
+  justify-items: stretch;
+  gap: 1rem;
+  padding: 1rem;
+  min-height: 100vh;
+}
+
+#aplose-skeleton-component h1,
+#aplose-skeleton-component #osmose-bar-component {
+  grid-column: 1 / 3;
+  justify-self: center;
+}
+
+#aplose-skeleton-component #nav {
+  display: grid;
+  align-content: baseline;
+  gap: 0.5rem;
+  padding: 1rem;
+}
+
+#aplose-skeleton-component #content {
+  padding: 1rem;
+}
